Split happiness.rs (529 lines) into happiness/ subdirectory
Split crates/simulation/src/happiness.rs into focused modules:
- constants.rs: coverage bitflags and happiness tuning constants
- coverage.rs: ServiceCoverageGrid type and update_service_coverage system
- systems.rs: HappinessExtras param and update_happiness system
- plugin.rs: HappinessPlugin registration
- tests.rs: unit tests
- mod.rs: thin facade with re-exports

All public items re-exported from mod.rs — no external import changes needed.
