+++ crates/simulation/src/happiness/mod.rs
@@ -1,0 +1,11 @@
+mod constants;
+mod coverage;
+mod plugin;
+mod systems;
+#[cfg(test)]
+mod tests;
+
+pub use constants::*;
+pub use coverage::{update_service_coverage, ServiceCoverageGrid};
+pub use plugin::HappinessPlugin;
+pub use systems::{update_happiness, HappinessExtras};
