+++ crates/simulation/src/happiness/systems.rs
@@ -1,7 +1,6 @@
 use bevy::prelude::*;
 
 use crate::citizen::{Citizen, CitizenDetails, HomeLocation, Needs, WorkLocation};
-use crate::config::{CELL_SIZE, GRID_HEIGHT, GRID_WIDTH};
 use crate::crime::CrimeGrid;
 use crate::death_care::{self, DeathCareGrid};
 use crate::economy::CityBudget;
@@ -10,187 +9,12 @@ use crate::heating::{self, HeatingGrid};
 use crate::homelessness::Homeless;
 use crate::policies::Policies;
 use crate::postal::PostalCoverage;
-use crate::services::{ServiceBuilding, ServiceType};
 use crate::traffic::TrafficGrid;
 use crate::wealth::WealthTier;
 use crate::weather::Weather;
 
-/// Bitflags for service coverage packed into a single byte per cell.
-pub const COVERAGE_HEALTH: u8 = 0b0000_0001;
-pub const COVERAGE_EDUCATION: u8 = 0b0000_0010;
-pub const COVERAGE_POLICE: u8 = 0b0000_0100;
-pub const COVERAGE_PARK: u8 = 0b0000_1000;
-pub const COVERAGE_ENTERTAINMENT: u8 = 0b0001_0000;
-pub const COVERAGE_TELECOM: u8 = 0b0010_0000;
-pub const COVERAGE_TRANSPORT: u8 = 0b0100_0000;
-pub const COVERAGE_FIRE: u8 = 0b1000_0000;
-
-/// Per-cell service coverage flags, precomputed when service buildings change.
-/// Uses bitflags packed into a single Vec<u8> — 5x less memory than 5 separate Vec<bool>.
-#[derive(Resource)]
-pub struct ServiceCoverageGrid {
-    /// One byte per cell, with bits for each service type.
-    pub flags: Vec<u8>,
-    pub dirty: bool,
-}
-
-impl Default for ServiceCoverageGrid {
-    fn default() -> Self {
-        let n = GRID_WIDTH * GRID_HEIGHT;
-        Self {
-            flags: vec![0; n],
-            dirty: true,
-        }
-    }
-}
-
-impl ServiceCoverageGrid {
-    pub fn clear(&mut self) {
-        self.flags.fill(0);
-    }
-
-    pub fn idx(x: usize, y: usize) -> usize {
-        y * GRID_WIDTH + x
-    }
-
-    #[inline]
-    pub fn has_health(&self, idx: usize) -> bool {
-        self.flags[idx] & COVERAGE_HEALTH != 0
-    }
-    #[inline]
-    pub fn has_education(&self, idx: usize) -> bool {
-        self.flags[idx] & COVERAGE_EDUCATION != 0
-    }
-    #[inline]
-    pub fn has_police(&self, idx: usize) -> bool {
-        self.flags[idx] & COVERAGE_POLICE != 0
-    }
-    #[inline]
-    pub fn has_park(&self, idx: usize) -> bool {
-        self.flags[idx] & COVERAGE_PARK != 0
-    }
-    #[inline]
-    pub fn has_entertainment(&self, idx: usize) -> bool {
-        self.flags[idx] & COVERAGE_ENTERTAINMENT != 0
-    }
-    #[inline]
-    pub fn has_telecom(&self, idx: usize) -> bool {
-        self.flags[idx] & COVERAGE_TELECOM != 0
-    }
-    #[inline]
-    pub fn has_transport(&self, idx: usize) -> bool {
-        self.flags[idx] & COVERAGE_TRANSPORT != 0
-    }
-    #[inline]
-    pub fn has_fire(&self, idx: usize) -> bool {
-        self.flags[idx] & COVERAGE_FIRE != 0
-    }
-}
-
-pub fn update_service_coverage(
-    services: Query<&ServiceBuilding>,
-    added_services: Query<Entity, Added<ServiceBuilding>>,
-    mut coverage: ResMut<ServiceCoverageGrid>,
-    ext_budget: Res<crate::budget::ExtendedBudget>,
-) {
-    if !added_services.is_empty() {
-        coverage.dirty = true;
-    }
-    if ext_budget.is_changed() {
-        coverage.dirty = true;
-    }
-    if !coverage.dirty {
-        return;
-    }
-    coverage.dirty = false;
-    coverage.clear();
-
-    for service in &services {
-        let budget_level = ext_budget.service_budgets.for_service(service.service_type);
-        let effective_radius = service.radius * budget_level;
-        let radius_cells = (effective_radius / CELL_SIZE).ceil() as i32;
-        let sx = service.grid_x as i32;
-        let sy = service.grid_y as i32;
-        let r2 = effective_radius * effective_radius;
-
-        // Determine which coverage bits this service sets
-        let bits = match service.service_type {
-            ServiceType::Hospital | ServiceType::MedicalClinic | ServiceType::MedicalCenter => {
-                COVERAGE_HEALTH
-            }
-            ServiceType::ElementarySchool
-            | ServiceType::HighSchool
-            | ServiceType::University
-            | ServiceType::Library
-            | ServiceType::Kindergarten => COVERAGE_EDUCATION,
-            ServiceType::PoliceStation
-            | ServiceType::PoliceKiosk
-            | ServiceType::PoliceHQ
-            | ServiceType::Prison => COVERAGE_POLICE,
-            ServiceType::SmallPark | ServiceType::LargePark | ServiceType::Playground => {
-                COVERAGE_PARK
-            }
-            ServiceType::Stadium | ServiceType::Plaza | ServiceType::SportsField => {
-                COVERAGE_ENTERTAINMENT
-            }
-            ServiceType::CellTower | ServiceType::DataCenter => COVERAGE_TELECOM,
-            ServiceType::BusDepot
-            | ServiceType::TrainStation
-            | ServiceType::SubwayStation
-            | ServiceType::TramDepot
-            | ServiceType::FerryPier
-            | ServiceType::SmallAirstrip
-            | ServiceType::RegionalAirport
-            | ServiceType::InternationalAirport => COVERAGE_TRANSPORT,
-            ServiceType::FireStation | ServiceType::FireHouse | ServiceType::FireHQ => {
-                COVERAGE_FIRE
-            }
-            _ => continue,
-        };
-
-        for dy in -radius_cells..=radius_cells {
-            for dx in -radius_cells..=radius_cells {
-                let cx = sx + dx;
-                let cy = sy + dy;
-                if cx < 0 || cy < 0 || cx >= GRID_WIDTH as i32 || cy >= GRID_HEIGHT as i32 {
-                    continue;
-                }
-                let wx_diff = dx as f32 * CELL_SIZE;
-                let wy_diff = dy as f32 * CELL_SIZE;
-                if wx_diff * wx_diff + wy_diff * wy_diff > r2 {
-                    continue;
-                }
-                let idx = ServiceCoverageGrid::idx(cx as usize, cy as usize);
-                coverage.flags[idx] |= bits;
-            }
-        }
-    }
-}
-
-const BASE_HAPPINESS: f32 = 50.0;
-const EMPLOYED_BONUS: f32 = 15.0;
-const SHORT_COMMUTE_BONUS: f32 = 10.0;
-const POWER_BONUS: f32 = 5.0;
-const NO_POWER_PENALTY: f32 = 25.0;
-const WATER_BONUS: f32 = 5.0;
-const NO_WATER_PENALTY: f32 = 20.0;
-const HEALTH_COVERAGE_BONUS: f32 = 5.0;
-const EDUCATION_BONUS: f32 = 3.0;
-const POLICE_BONUS: f32 = 5.0;
-const PARK_BONUS: f32 = 8.0;
-const ENTERTAINMENT_BONUS: f32 = 5.0;
-const HIGH_TAX_PENALTY: f32 = 8.0;
-const CONGESTION_PENALTY: f32 = 5.0;
-const GARBAGE_PENALTY: f32 = 5.0;
-const CRIME_PENALTY_MAX: f32 = 15.0;
-const TELECOM_BONUS: f32 = 3.0;
-const TRANSPORT_BONUS: f32 = 4.0;
-const POOR_ROAD_PENALTY: f32 = 3.0;
-
-/// Happiness penalty for homeless citizens (unsheltered).
-pub const HOMELESS_PENALTY: f32 = 30.0;
-/// Reduced happiness penalty for sheltered homeless citizens.
-pub const SHELTERED_PENALTY: f32 = 10.0;
+use super::constants::*;
+use super::coverage::ServiceCoverageGrid;
 
 /// Bundled secondary resources for update_happiness to stay within the 16-param limit.
 #[derive(bevy::ecs::system::SystemParam)]
@@ -422,108 +246,4 @@ pub fn update_happiness(
 
             details.happiness = happiness.clamp(0.0, 100.0);
         });
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_happiness_bounds() {
-        // Base happiness should be in range
-        assert!(BASE_HAPPINESS >= 0.0 && BASE_HAPPINESS <= 100.0);
-    }
-
-    #[test]
-    fn test_all_factors_affect_output() {
-        // Verify all bonuses/penalties are non-zero
-        assert!(EMPLOYED_BONUS > 0.0);
-        assert!(SHORT_COMMUTE_BONUS > 0.0);
-        assert!(POWER_BONUS > 0.0);
-        assert!(WATER_BONUS > 0.0);
-        assert!(HEALTH_COVERAGE_BONUS > 0.0);
-        assert!(EDUCATION_BONUS > 0.0);
-        assert!(POLICE_BONUS > 0.0);
-        assert!(PARK_BONUS > 0.0);
-        assert!(ENTERTAINMENT_BONUS > 0.0);
-        assert!(HIGH_TAX_PENALTY > 0.0);
-        assert!(CONGESTION_PENALTY > 0.0);
-        assert!(GARBAGE_PENALTY > 0.0);
-    }
-
-    #[test]
-    fn test_max_happiness_reachable() {
-        // Max theoretical happiness: all bonuses, no penalties, max land value (255/50 = 5.1)
-        let max_land_bonus: f32 = 255.0 / 50.0;
-        let max = BASE_HAPPINESS
-            + EMPLOYED_BONUS
-            + SHORT_COMMUTE_BONUS
-            + POWER_BONUS
-            + WATER_BONUS
-            + HEALTH_COVERAGE_BONUS
-            + EDUCATION_BONUS
-            + POLICE_BONUS
-            + PARK_BONUS
-            + ENTERTAINMENT_BONUS
-            + max_land_bonus;
-        // With all bonuses the raw sum exceeds 100, but clamp caps at 100
-        assert!(
-            max > 100.0,
-            "max happiness {} should exceed 100 before clamping",
-            max
-        );
-        // Verify it is meaningful without land value
-        let max_no_land = BASE_HAPPINESS
-            + EMPLOYED_BONUS
-            + SHORT_COMMUTE_BONUS
-            + POWER_BONUS
-            + WATER_BONUS
-            + HEALTH_COVERAGE_BONUS
-            + EDUCATION_BONUS
-            + POLICE_BONUS
-            + PARK_BONUS
-            + ENTERTAINMENT_BONUS;
-        assert!(
-            max_no_land > 80.0,
-            "max happiness {} (no land) should be > 80 to be meaningful",
-            max_no_land
-        );
-    }
-
-    #[test]
-    fn test_service_coverage_dirty_flag_default() {
-        let grid = ServiceCoverageGrid::default();
-        assert!(grid.dirty, "should start dirty so first update runs");
-    }
-
-    #[test]
-    fn test_service_coverage_clear_resets_all() {
-        let mut grid = ServiceCoverageGrid::default();
-        let idx = ServiceCoverageGrid::idx(10, 10);
-        grid.flags[idx] = COVERAGE_HEALTH
-            | COVERAGE_EDUCATION
-            | COVERAGE_POLICE
-            | COVERAGE_PARK
-            | COVERAGE_ENTERTAINMENT;
-        grid.clear();
-        assert!(!grid.has_health(idx));
-        assert!(!grid.has_education(idx));
-        assert!(!grid.has_police(idx));
-        assert!(!grid.has_park(idx));
-        assert!(!grid.has_entertainment(idx));
-    }
-}
-
-pub struct HappinessPlugin;
-
-impl Plugin for HappinessPlugin {
-    fn build(&self, app: &mut App) {
-        app.init_resource::<ServiceCoverageGrid>().add_systems(
-            FixedUpdate,
-            (update_service_coverage, update_happiness)
-                .chain()
-                .after(crate::postal::update_postal_coverage)
-                .in_set(crate::SimulationSet::Simulation),
-        );
-    }
 }
